Adding new project interface. Moving to upload.
New idea for components; Some components have several pieces. An empty
component could be placed somewhere for a second one to use as a
container. For example, a date picker search field could use an empty
component to display the date layout wherever rather than constrained
in place. Other components display could be switched off when the date
field is focused as well.